Updating archivist to reflect storage reality
Archivist doesn't quite behave as specified in it's help text, the help text
and behaviour have both been corrected. This includes changing archivist to
correctly notice if the archive is missing and doesn't contain any entries.

This also updates archivist to use a different search strategy for it's storage
files. It now checks for user input "-a archive" then checks the file system
for a ".archivist" directory, and will finally resort to using the
"~/.config/archivist.d/" directory.

WARNING: This is a breaking change, archivist no longer uses "~/.archivist" as
it's default storage location.