dot-init verifies that source/dest files differ
dot-init now checks that the source and dest file are different. It
is possible that they were the same file, a symbolic link would do
this, and checking for a symbolic link is insufficient. As an example:
a directory could be linked to another, making files inside the same,
but the files wouldn't be links.