Vim subshell history based on vim pid
Basing the vim history on the pid of the vim process that spawned
the bash shell makes for much nicer history, since it's more likely
that I'm going to do 'edit, sh, edit, sh, ...'.

Additionally, removing the current time from the history file makes
for a much easier time finding the files.