Adds time-until
time-until is a small command that wraps date(1) to provide
a simple interface to time duration math. It accepts 3 args:
    ./time-until <time-spec> ["in" <unit>]
where <time-spec> is something understood by date(1)'s --date
parameter, and <unit> is a time unit seconds..years, it
defaults to seconds. time-until outputs the number of <unit>s
until or since (negative) the specified time, as either
integers or decimals to two places. The "in" and <unit> are
optional, but must be provided as a pair, if present.