Adds chrome-i3
Chrome isn't part of my i3 startup configuration, and gets started
later. This means that I make an attempt to position it on a specific
i3 workspace, and want it to stay there. This doesn't work on the
first startup with cold cache, because chrome takes just long enough
to start that I've context switched to something else while I wait.

chrome-i3 wraps starting chrome with an i3-msg command to put in the
placeholder i3 window. This causes chrome to start where it was
originally spawned (assuming it only creates one window) and not pop
up on whatever workspace is currently focused.