Updates now timeout strategy
Uses sleep to correctly get the exit value of read when it times out.

Also avoids file descriptor shenanigans, which cause problems when
setting the output of the script to be a pipeline instead of a
terminal. eg: `... | now | tee`

Terminal (tty) file descriptors don't really (as far as I can tell)
differentiate between input or output, so you can read from a
descriptor when it's passed to you for writing to (as, eg: stdout).
You shouldn't rely on this behaviour, because someone could stop
passing you a terminal, and instead give you a fifo (eg: this case).