Adds steam wrapper
Steams runs 'fossilize_replay' to pre-compile game textures. For some
reason this binary eats all the CPU & RAM and causes my systems die
because of "out of memory" errors. That is: way too much memory was
allocated and the kernel physically can not produce that much.

This is fixed in 3 parts:
 * steam - steam wrapper
 * steam-fake-game - fake "game" to run through steam
 * steam-fossilize-wrap - wrapper for fossilize_replay

The way this works:

 steam and steam-fake-game are used to start steam (wrapped) and
 then detect when steam has finished running updates and
 bootstrapping. Once we know steam is ready to run games, the steam
 wrapper wraps fossilize_replay with our wrapper, which limits the
 amount of ram, cpu, and io it can use.

This depends on `pubsub-fifo` and `same-file`, but should otherwise
run on a basic linux install?

See also https://github.com/ValveSoftware/Fossilize/issues/84