Reimplement pubsub-fifo
Switch to using unix domain sockets instead of fifos. These turn out
to behave much more reliably when the reader end is in the process
of shutting down.

With fifos there's a race during shutdown, because the fifo is not
yet unlinked and can still be opened for writing, except that it
will not be opened again for reading. If a publisher manages to open
it just before unlink, then it will never be read from again, and the
writer will wait forever for a reader to empty the fifo.

With unix sockets a reader is required before a publisher can open
the socket. During shutdown no writers are accepted and attempts to
open the socket all fail. This means that no writer gets stuck in a
weird state.