Properly handle napping
Previously the expectation was that duplicate entries for the same
day were the result of the export from Sleepbot. This is partially
true, because Sleepbot does output duplicate lines at times, but
duplicate dates are also caused by naps.

Removing duplicate lines cleans up the output data, and allows naps
to be output.