Add SO_BINDTODEVICE functionality
On linux SO_BINDTODEVICE allows us to bind to a specific network
interface instead of possibly binding to everywhere or an ip address.
This enables, for example, binding to a VPN only, or only localhost.
There's some finagling to support passing either an IP/hostname or an
interface name via the argument.